backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Google Cloud
google-cloud-storage>=2.14.0
google-cloud-firestore>=2.14.0

# AI Model SDKs
openai>=1.50.0
google-generativeai>=0.8.0
# dashscope>=1.20.0  # Alibaba Tongyi (uncomment when needed)
# volcengine-python-sdk  # ByteDance Doubao (uncomment when needed)

# Development
httpx>=0.27.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0

# Auth
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9


